fix(ci): repair the jscpd baseline and clear the checks it hid
Why

`.jscpd-baseline.json` lost a closing brace in #2511, so every `bun run
verify` on main since has failed twice over: the jscpd ratchet crashed
parsing its own baseline, and `prettier-root` could not parse the file
either. With the ratchet down, real duplication regressions landed
unnoticed. Two more lanes were red alongside it: `bake-images.test.ts`
had grown past the 1500-line cap, and #2590 narrowed the Scout detached
work `kind` union to a single member without updating the test and two
guards that still discriminate on it.

What

- restore the missing `},` in `.jscpd-baseline.json` so both lanes parse
- deduplicate what the ratchet had stopped catching: extract a
  `temporalFailureWebhook` fixture for the three copies of the
  `TemporalWorkflowFailed` webhook builder, move both hand-rolled argv
  parsers onto `node:util` `parseArgs` (the repo idiom), and fold the
  workflow-failure-watch tests onto the existing
  `captureFirstFailureDescription` helper plus new `closedExecutions` /
  `clientForExecutions` / `failListAfterCurrentPage` builders — 296
  duplication warnings down to 20 in that file
- tighten the baseline to the now-lower counts (it only moves one way)
- split `bake-images.test.ts` (1569 lines) by module: the Temporal
  admission tests become `temporal-candidate-admission.test.ts` and the
  `pinCandidatesForDigests` tests join the existing
  `pin-candidate-images.test.ts`
- finish #2590's removal: point the work-store integration test at
  `parlay-generation` and drop the two now-vacuous `kind` comparisons

Verification

- bun run jscpd (ratchet passes; baseline diff only lowers counts)
- bun run prettier-root
- bunx turbo run typecheck lint test --filter=@shepherdjerred/temporal
  --filter=@shepherdjerred/root-scripts --filter=@scout-for-lol/backend
  --filter=@shepherdjerred/alert-dashboard
- DATABASE_URL=file:… bun run --cwd packages/alert-dashboard test:sqlite
  (14 tests)
- bunx lefthook run pre-commit
