+++ .buildkite/scripts/pin-candidate-images.test.ts
@@ -205,3 +205,61 @@ test("does not publish a Scout beta pin while tracks diverge", () => {
     "shepherdjerred/scout-for-lol/beta": { version: "2.0.0-44", digest },
   });
 });
+
+test("retains a central Workflow candidate until its pin converges with stable", () => {
+  const digest = `sha256:${"b".repeat(64)}`;
+  expect(
+    pinCandidatesForDigests(
+      { "shepherdjerred/temporal-worker": digest },
+      "44",
+      versionCatalogSource([
+        {
+          name: "shepherdjerred/temporal-worker/workflows/candidate",
+          value: `2.0.0-43@sha256:${"c".repeat(64)}`,
+        },
+        {
+          name: "shepherdjerred/temporal-worker/workflows/stable",
+          value: `2.0.0-42@sha256:${"d".repeat(64)}`,
+        },
+      ]),
+    ),
+  ).toEqual({
+    "shepherdjerred/temporal-worker": {
+      version: "2.0.0-44",
+      digest,
+    },
+  });
+});
+test("bootstraps the legacy Scout beta stable and candidate pins", () => {
+  const digest = `sha256:${"b".repeat(64)}`;
+  const legacy = `2.0.0-12197@sha256:${"c".repeat(64)}`;
+  expect(
+    pinCandidatesForDigests(
+      { "shepherdjerred/scout-for-lol/beta": digest },
+      "43",
+      versionCatalogSource([
+        {
+          name: "shepherdjerred/scout-for-lol/beta/workflows/candidate",
+          value: legacy,
+        },
+        {
+          name: "shepherdjerred/scout-for-lol/beta/workflows/stable",
+          value: legacy,
+        },
+      ]),
+    ),
+  ).toEqual({
+    "shepherdjerred/scout-for-lol/beta": {
+      version: "2.0.0-43",
+      digest,
+    },
+    "shepherdjerred/scout-for-lol/beta/workflows/stable": {
+      version: "2.0.0-43",
+      digest,
+    },
+    "shepherdjerred/scout-for-lol/beta/workflows/candidate": {
+      version: "2.0.0-43",
+      digest,
+    },
+  });
+});
